# Conflicts:
#	air_ticket_booking/src/main/java/com/example/air_ticket_booking/controller/route/RouteController.java
#	air_ticket_booking/src/main/java/com/example/air_ticket_booking/repository/route/IRouteRepository.java

+++ air_ticket_booking/src/main/java/com/example/air_ticket_booking/controller/ticket/TicketController.java
@@ -6,11 +6,11 @@ import com.example.air_ticket_booking.repository.ticket.ITicketRepository;
 import com.example.air_ticket_booking.service.ticket.ITicketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
-
 import org.springframework.web.bind.annotation.*;
 import com.example.air_ticket_booking.dto.ticket.TicketDto;
 import com.example.air_ticket_booking.model.customer.Customer;
@@ -19,14 +19,9 @@ import com.example.air_ticket_booking.model.seat.Seat;
 import com.example.air_ticket_booking.model.ticket.Ticket;
 import com.example.air_ticket_booking.model.ticket.TypeTicket;
 import com.example.air_ticket_booking.model.type_passenger.TypePassenger;
-import com.example.air_ticket_booking.service.ticket.ITicketService;
-import com.example.air_ticket_booking.service.ticket.impl.TicketService;
-import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+
+import java.util.Arrays;
 
 @RestController
 @CrossOrigin("*")
@@ -129,13 +124,21 @@ public class TicketController {
     /**
      * task response data ticket booked to FE
      * @Method showAllTickets
-     * @param pageable
+     * @param page
      * @return HttpStatus and Page<Ticket>
      * @author Nhàn NA
      */
-    @GetMapping()
-    public ResponseEntity<Page<ITicketProjection>> showAllTickets(Pageable pageable){
-        System.out.println("nhan");
+    @GetMapping("/booked/{page}")
+    public ResponseEntity<Page<ITicketProjection>> showAllTickets(@PathVariable int page){
+        String idString= String.valueOf(page);
+        if(!idString.matches("^[0-9]{1,8}$")){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        if(page<0){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        Pageable pageable= PageRequest.of(page,5);
+
         if(iTicketService.findAllTickets(pageable).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }else {
@@ -146,12 +149,17 @@ public class TicketController {
     /**
      * task delete ticket value id
      * @Method deleteTicket
+     * date create: 10/08/2023
      * @param id
      * @return HttpStatus
      * @author Nhàn NA
      */
     @DeleteMapping("{id}")
     public ResponseEntity<?> deleteTicket(@PathVariable Long id){
+        String idString= String.valueOf(id);
+        if(!idString.matches("^[0-9]{1,8}$")){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         if(iTicketService.deleteTicket(id)) {
             return new ResponseEntity<>(HttpStatus.OK);
         }else {
@@ -161,13 +169,20 @@ public class TicketController {
     /**
      * task response search all ticket booked data to FE
      * @Method searchTickets
+     * date create: 10/08/2023
      * @param item,pageable
      * @return HttpStatus and Page<Ticket>
      * @author Nhàn NA
      */
-    @GetMapping("/search/{item}")
-    public ResponseEntity<Page<ITicketProjection>> searchTickets(@PathVariable String item,Pageable pageable){
-        String[] input = item.split(",");
+    @GetMapping("/search/{item}/{page}")
+    public ResponseEntity<Page<ITicketProjection>> searchTickets(@PathVariable("item") String item,@PathVariable("page") int page){
+        String idString= String.valueOf(page);
+        String[] input = item.split(",", -1);
+        System.out.println(Arrays.toString(input));
+        Pageable pageable = PageRequest.of(page,5);
+        if(!input[0].matches("^[0-9]{1,8}$")||input.length!=5||!idString.matches("^[0-9]{1,8}$")){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         if(iTicketService.searchTicket(Long.valueOf(input[0]),input[1],input[2],input[3],input[4],pageable).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
@@ -178,13 +193,18 @@ public class TicketController {
     /**
      * task response ticket unbooked  data to FE
      * @Method findAllTicketUnbooked
-     * @param pageable
+     * date create: 10/08/2023
+     * @param page
      * @return HttpStatus and Page<Ticket>
      * @author Nhàn NA
      */
-    @GetMapping("/unbooked")
-    public ResponseEntity<Page<ITicketUnbookedProjection>> findAllTicketUnbooked(Pageable pageable){
-        System.out.println("nhan");
+    @GetMapping("/unbooked/{page}")
+    public ResponseEntity<Page<ITicketUnbookedProjection>> findAllTicketUnbooked(@PathVariable int page){
+        String idString= String.valueOf(page);
+        if(page<0||!idString.matches("^[0-9]{1,8}$")){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        Pageable pageable=PageRequest.of(page,5);
         if(iTicketService.findAllTicketUnbooked(pageable).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }else {
@@ -194,13 +214,19 @@ public class TicketController {
     /**
      * task response search unbooked tickets   data to FE
      * @Method findAllTicketUnbooked
-     * @param pageable
+     * date create: 10/08/2023
+     * @param item,page
      * @return HttpStatus and Page<Ticket>
      * @author Nhàn NA
      */
-    @GetMapping("/search-unbooked/{item}")
-    public ResponseEntity<Page<ITicketUnbookedProjection>> searchTicketsUnBooked(@PathVariable String item,Pageable pageable){
-        String[] input = item.split(",");
+    @GetMapping("/search-unbooked/{item}/{page}")
+    public ResponseEntity<Page<ITicketUnbookedProjection>> searchTicketsUnBooked(@PathVariable("item") String item, @PathVariable("page") int page){
+        String[] input = item.split(",", -1);
+        Pageable pageable= PageRequest.of(page,5);
+        String idString= String.valueOf(page);
+        if(page<0||!idString.matches("^[0-9]{1,8}$")||!input[0].matches("^[0-9]{1,8}$")){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         if(iTicketService.searchTicketUnbooked(Long.valueOf(input[0]),input[1],input[2],input[3],input[4],pageable).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
